Resolve #29 centralize event name matching/normalising and use it in both the event sync strategy and the fanduel controller
