+++ backend/src/golf/strategies/golf-event-sync.strategy.ts
@@ -2,15 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { FanduelService } from '@/external-providers/fanduel/fanduel.service';
 import { EspnService } from '@/external-providers/espn/espn.service';
 import { IEventSyncStrategy, EventSyncGroup } from '@/events/strategies/event-sync-strategy.interface';
-import { normalizeEventName } from '@/golf/golf-scoring.util';
-
-/**
- * Maps normalised FanDuel event names to their ESPN equivalents.
- * Keys and values must be in normalised form (lowercase, no punctuation).
- */
-const GOLF_EVENT_NAME_ALIASES: Record<string, string> = {
-  'the open championship': 'the open',
-};
+import { matchesEvent } from '@/golf/golf.util';
 
 @Injectable()
 export class GolfEventSyncStrategy implements IEventSyncStrategy {
@@ -29,7 +21,7 @@ export class GolfEventSyncStrategy implements IEventSyncStrategy {
     return fanduelEvents
       .map((fanduelEvent) => {
         const espnMatch = espnSchedule.find((espn) =>
-          this.matchesEvent(fanduelEvent.name, espn.name),
+          matchesEvent(fanduelEvent.name, espn.name),
         );
         if (!espnMatch) return null;
 
@@ -47,14 +39,5 @@ export class GolfEventSyncStrategy implements IEventSyncStrategy {
         };
       })
       .filter((group): group is EventSyncGroup => group !== null);
-  }
-
-  private matchesEvent(fanduelName: string, espnName: string): boolean {
-    const fdNormalised = normalizeEventName(fanduelName);
-    const espnNormalised = normalizeEventName(espnName);
-
-    const fdMatch = GOLF_EVENT_NAME_ALIASES[fdNormalised] ?? fdNormalised;
-
-    return espnNormalised.includes(fdMatch) || fdMatch.includes(espnNormalised);
   }
 }
